backend/services/media_service/requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
redis==5.0.1

# Celery for async tasks
celery[redis]==5.3.4

# HTTP client
httpx==0.26.0
requests==2.31.0

# Image processing
Pillow==10.2.0

# Cloud storage
# supabase==2.3.0  # If using Supabase
boto3==1.34.34  # AWS S3
botocore==1.34.34

# Utilities
python-dotenv==1.0.0
python-json-logger==2.0.7
